+++ core-mvp-binding/RELEASE_NOTES.md
@@ -5,9 +5,9 @@
 * ANDDEP-687 Added "androidx.constraintlayout:constraintlayout" dependency with "implementation" type
 * Clearing `BaseRxFragmentView.viewDisposable` move from onDestroy to onDestroyView
 * ANDDEP-671 Renamed loadable to response
+* AndroidSchedulers.mainThread() replaced with MainThreadImmediateScheduler
+ in subscription methods of BaseRxViews.
 * Fixed RequestState's observeHasError method + added observeOptionalData
 ## 0.3.0
 ##### Core-mvp-binding
 * Renamed `onViewDetached ()` -> `onViewDetach ()`
-* AndroidSchedulers.mainThread() replaced with ImmediateScheduler 
- in subscription methods of BaseRxViews.